// src/app/@core/models/bill.model.ts
export interface Bill {
  electric_number: number;
  id: number;
  month: number;
  network_number: number;
  room_name: string;
  room_price: number;
  status: string;
  store_name: string;
  total_bill: number;
  total_electric_price: number;
  total_network_price: number;
  total_water_price: number;
  username: string;
  water_number: number;
  year: number;
}